+++ src/pypi_attestations/_impl.py
@@ -342,6 +342,11 @@ class Envelope(BaseModel):
     """
 
 
+def _der_decode_utf8string(der: bytes) -> str:
+    """Decode a DER-encoded UTF8String."""
+    return der_decode(der, UTF8String)[0].decode()  # type: ignore[no-any-return]
+
+
 def _ultranormalize_dist_filename(dist: str) -> str:
     """Return an "ultranormalized" form of the given distribution filename.
 
@@ -424,11 +429,6 @@ class _GitHubTrustedPublisherPolicy:
             ]
         )
 
-    @classmethod
-    def _der_decode_utf8string(cls, der: bytes) -> str:
-        """Decode a DER-encoded UTF8String."""
-        return der_decode(der, UTF8String)[0].decode()  # type: ignore[no-any-return]
-
     def verify(self, cert: Certificate) -> None:
         """Verify the certificate against the Trusted Publisher identity."""
         self._subpolicy.verify(cert)
@@ -452,18 +452,18 @@ class _GitHubTrustedPublisherPolicy:
         #  where OWNER/REPO and WORKFLOW are controlled by the TP identity,
         #  and REF is controlled by the certificate's own claims.
         build_config_uri = cert.extensions.get_extension_for_oid(policy._OIDC_BUILD_CONFIG_URI_OID)  # noqa: SLF001
-        raw_build_config_uri = self._der_decode_utf8string(build_config_uri.value.public_bytes())
+        raw_build_config_uri = _der_decode_utf8string(build_config_uri.value.public_bytes())
 
         # (2) Extract the source repo digest and ref.
         source_repo_digest = cert.extensions.get_extension_for_oid(
             policy._OIDC_SOURCE_REPOSITORY_DIGEST_OID  # noqa: SLF001
         )
-        sha = self._der_decode_utf8string(source_repo_digest.value.public_bytes())
+        sha = _der_decode_utf8string(source_repo_digest.value.public_bytes())
 
         source_repo_ref = cert.extensions.get_extension_for_oid(
             policy._OIDC_SOURCE_REPOSITORY_REF_OID  # noqa: SLF001
         )
-        ref = self._der_decode_utf8string(source_repo_ref.value.public_bytes())
+        ref = _der_decode_utf8string(source_repo_ref.value.public_bytes())
 
         # (3)-(4): Build the expected URIs and compare them
         for suffix in [sha, ref]:
@@ -507,6 +507,71 @@ class GitHubPublisher(_PublisherBase):
         return _GitHubTrustedPublisherPolicy(self.repository, self.workflow)
 
 
+class _GitLabTrustedPublisherPolicy:
+    """A custom sigstore-python policy for verifying against a GitLab-based Trusted Publisher."""
+
+    def __init__(self, repository: str, workflow_filepath: str) -> None:
+        self._repository = repository
+        self._workflow_filepath = workflow_filepath
+        # This policy must also satisfy some baseline underlying policies:
+        # the issuer must be GitHub Actions, and the repo must be the one
+        # we expect.
+        self._subpolicy = policy.AllOf(
+            [
+                policy.OIDCIssuerV2("https://gitlab.com"),
+                policy.OIDCSourceRepositoryURI(f"https://gitlab.com/{self._repository}"),
+            ]
+        )
+
+    def verify(self, cert: Certificate) -> None:
+        """Verify the certificate against the Trusted Publisher identity."""
+        self._subpolicy.verify(cert)
+
+        # This process has a few annoying steps, since a Trusted Publisher
+        # isn't aware of the commit or ref it runs on, while Sigstore's
+        # leaf certificate claims (like GitLab CI/CD's OIDC claims) only
+        # ever encode the workflow filename (which we need to check) next
+        # to the ref/sha (which we can't check).
+        #
+        # To get around this, we:
+        # (1) extract the `Build Config URI` extension;
+        # (2) extract the `Source Repository Digest` and
+        #     `Source Repository Ref` extensions;
+        # (3) build the *expected* URI with the user-controlled
+        #     Trusted Publisher identity *with* (2)
+        # (4) compare (1) with (3)
+
+        # (1) Extract the build config URI, which looks like this:
+        #     https://gitlab.com/NAMESPACE/PROJECT//WORKFLOW_FILEPATH@REF
+        #  where NAMESPACE/PROJECT and WORKFLOW_FILEPATH are controlled by the TP identity,
+        #  and REF is controlled by the certificate's own claims.
+        build_config_uri = cert.extensions.get_extension_for_oid(policy._OIDC_BUILD_CONFIG_URI_OID)  # noqa: SLF001
+        raw_build_config_uri = _der_decode_utf8string(build_config_uri.value.public_bytes())
+
+        # (2) Extract the source repo digest and ref.
+        source_repo_digest = cert.extensions.get_extension_for_oid(
+            policy._OIDC_SOURCE_REPOSITORY_DIGEST_OID  # noqa: SLF001
+        )
+        sha = _der_decode_utf8string(source_repo_digest.value.public_bytes())
+
+        source_repo_ref = cert.extensions.get_extension_for_oid(
+            policy._OIDC_SOURCE_REPOSITORY_REF_OID  # noqa: SLF001
+        )
+        ref = _der_decode_utf8string(source_repo_ref.value.public_bytes())
+
+        # (3)-(4): Build the expected URIs and compare them
+        for suffix in [sha, ref]:
+            expected = f"https://gitlab.com/{self._repository}//{self._workflow_filepath}@{suffix}"
+            if raw_build_config_uri == expected:
+                return
+
+        # If none of the expected URIs matched, the policy fails.
+        raise sigstore.errors.VerificationError(
+            f"Certificate's Build Config URI ({build_config_uri}) does not match expected "
+            f"Trusted Publisher ({self._workflow_filepath} @ {self._repository})"
+        )
+
+
 class GitLabPublisher(_PublisherBase):
     """A GitLab-based Trusted Publisher."""
 
@@ -519,30 +584,19 @@ class GitLabPublisher(_PublisherBase):
     `bar` owned by group `foo` and subgroup `baz`.
     """
 
+    workflow_filepath: str
+    """
+    The path for the CI/CD configuration file. This is usually ".gitlab-ci.yml",
+    but can be customized.
+    """
+
     environment: Optional[str] = None
     """
     The optional environment that the publishing action was performed from.
     """
 
     def _as_policy(self) -> VerificationPolicy:
-        policies: list[VerificationPolicy] = [
-            policy.OIDCIssuerV2("https://gitlab.com"),
-            policy.OIDCSourceRepositoryURI(f"https://gitlab.com/{self.repository}"),
-        ]
-
-        if not self.claims:
-            raise VerificationError("refusing to build a policy without claims")
-
-        if ref := self.claims.get("ref"):
-            policies.append(
-                policy.OIDCBuildConfigURI(
-                    f"https://gitlab.com/{self.repository}//.gitlab-ci.yml@{ref}"
-                )
-            )
-        else:
-            raise VerificationError("refusing to build a policy without a ref claim")
-
-        return policy.AllOf(policies)
+        return _GitLabTrustedPublisherPolicy(self.repository, self.workflow_filepath)
 
 
 _Publisher = Union[GitHubPublisher, GitLabPublisher]