+++ test/test_impl.py
@@ -35,6 +35,10 @@ gh_signed_dist_path = _ASSETS / "pypi_attestation_models-0.0.4a2.tar.gz"
 gh_signed_dist = impl.Distribution.from_file(gh_signed_dist_path)
 gh_signed_dist_bundle_path = _ASSETS / "pypi_attestation_models-0.0.4a2.tar.gz.sigstore"
 
+gl_signed_dist_path = _ASSETS / "gitlab_oidc_project-0.0.3.tar.gz"
+gl_signed_dist = impl.Distribution.from_file(gl_signed_dist_path)
+gl_attestation_path = _ASSETS / "gitlab_oidc_project-0.0.3.tar.gz.publish.attestation"
+
 
 class TestDistribution:
     def test_from_file_nonexistent(self, tmp_path: Path) -> None:
@@ -147,6 +151,17 @@ class TestAttestation:
         assert predicate_type == "https://docs.pypi.org/attestations/publish/v1"
         assert predicate == {}
 
+    def test_verify_from_gitlab_publisher(self) -> None:
+        publisher = impl.GitLabPublisher(
+            repository="facutuesca/gitlab-oidc-project",
+            workflow_filepath=".gitlab-ci.yml",
+        )
+
+        attestation = impl.Attestation.model_validate_json(gl_attestation_path.read_text())
+        predicate_type, predicate = attestation.verify(publisher, gl_signed_dist)
+        assert predicate_type == "https://docs.pypi.org/attestations/publish/v1"
+        assert predicate is None
+
     def test_verify_from_github_publisher_wrong(self) -> None:
         publisher = impl.GitHubPublisher(
             repository="trailofbits/pypi-attestation-models",
@@ -158,6 +173,16 @@ class TestAttestation:
 
         with pytest.raises(impl.VerificationError, match=r"Build Config URI .+ does not match"):
             attestation.verify(publisher, gh_signed_dist)
+
+    def test_verify_from_gitlab_publisher_wrong(self) -> None:
+        publisher = impl.GitLabPublisher(
+            repository="facutuesca/gitlab-oidc-project",
+            workflow_filepath="wrong.yml",
+        )
+
+        attestation = impl.Attestation.model_validate_json(gl_attestation_path.read_text())
+        with pytest.raises(impl.VerificationError, match=r"Build Config URI .+ does not match"):
+            attestation.verify(publisher, gl_signed_dist)
 
     def test_verify(self) -> None:
         # Our checked-in asset has this identity.
@@ -541,10 +566,16 @@ class TestPublisher:
         assert gh.workflow == "publish.yml"
         assert TypeAdapter(impl.Publisher).validate_json(json.dumps(gh_raw)) == gh
 
-        gl_raw = {"kind": "GitLab", "repository": "foo/bar/baz", "environment": "publish"}
+        gl_raw = {
+            "kind": "GitLab",
+            "repository": "foo/bar/baz",
+            "workflow_filepath": "dir/release.yml",
+            "environment": "publish",
+        }
         gl: impl.Publisher = TypeAdapter(impl.Publisher).validate_python(gl_raw)
         assert isinstance(gl, impl.GitLabPublisher)
         assert gl.repository == "foo/bar/baz"
+        assert gl.workflow_filepath == "dir/release.yml"
         assert gl.environment == "publish"
         assert TypeAdapter(impl.Publisher).validate_json(json.dumps(gl_raw)) == gl
 
@@ -571,21 +602,6 @@ class TestPublisher:
             "this": "is-preserved",
             "this-too": 123,
         }
-
-
-class TestGitLabPublisher:
-    def test_as_policy(self) -> None:
-        publisher = impl.GitLabPublisher(repository="fake/fake", claims={"ref": "refs/heads/main"})
-        pol: policy.AllOf = publisher._as_policy()  # type: ignore[assignment]
-
-        assert len(pol._children) == 3
-
-    @pytest.mark.parametrize("claims", [None, {}, {"something": "unrelated"}, {"ref": None}])
-    def test_as_policy_invalid(self, claims: Optional[dict]) -> None:
-        publisher = impl.GitLabPublisher(repository="fake/fake", claims=claims)
-
-        with pytest.raises(impl.VerificationError, match="refusing to build a policy"):
-            publisher._as_policy()
 
 
 class TestProvenance: