+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- The `GitLabPublisher` policy now takes the workflow file path in order to
+  verify attestations, rathen than assuming it will always be `gitlab-ci.yml`
+  ([#71](https://github.com/trailofbits/pypi-attestations/pull/71)).
+- The `GitLabPublisher` now longer expects claims being passed during construction,
+  rather the `ref` and `sha` claims are extracted from the certificate's extensions,
+  similar to `GitHubPublisher`'s behavior
+  ([#71](https://github.com/trailofbits/pypi-attestations/pull/71)).
+
 ## [0.0.16]
 
 ### Added